added song queue into game scene screen, styling needed
