+++ src/client/Components/Spotify_Components/index.tsx
@@ -1,3 +1,4 @@
 import { SongSelection } from "./SongSelection/SongSelection";
+import { InGameSongQueue } from "./GameQueueComponents/InGameSongQueue";
 
-export { SongSelection };
+export { SongSelection, InGameSongQueue };
